.github/workflows/requirements.txt
torch==2.0.1
torchvision==0.15.2
realesrgan==0.3.0
basicsr==1.4.2
Pillow>=10.0.0
opencv-python-headless>=4.8.0
tqdm>=4.65.0
numpy>=1.24.0